# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cottage Village API Documentation",
        "description": "",
        "contact": {
            "email": "afonico@info.ru"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/admin/verify/user": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Verify user account. Accessable only by administator users",
                "description": "",
                "operationId": "verify",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "user_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "description": "",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/booking/all/get": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get all users bookings",
                "description": "",
                "operationId": "getAllBookings",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "service_id": {
                                        "description": "",
                                        "type": "integer"
                                    },
                                    "user_id": {
                                        "description": "",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get a JWT via given credentials",
                "description": "",
                "operationId": "login",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Users email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Users password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Return token object { 'access_token', 'token_type', 'expires_in' }"
                    }
                }
            }
        },
        "/api/auth/me": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get the authenticated User",
                "description": "",
                "operationId": "me",
                "responses": {
                    "200": {
                        "description": "Returns user data"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Log the user out (Invalidate the token)",
                "description": "",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/auth/refresh": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Refresh a token",
                "description": "",
                "operationId": "refresh",
                "responses": {
                    "200": {
                        "description": "Return token object { 'access_token', 'token_type', 'expires_in' }"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format: token_type access_token",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}